implemented leads retrieve service

+++ src/main/java/com/addi/test/leads_checker/domain/LeadsService.java
@@ -3,5 +3,5 @@ package com.addi.test.leads_checker.domain;
 import java.util.Optional;
 
 public interface LeadsService {
-    Optional<Lead> findById(Integer id);
+    Optional<Lead> findById(Long id);
 }
